MPEG-1307 criando metodo
